// ClientApp/app/models/perfil-profissional.ts
export class PerfilProfissional {
    cargo: string;
    cidade: string;
    bairro: string;
    complemento: string;
    dsProfissao: string;
    empresa: string;
    estado: string;
    idEstado: string;
    idCargo: number;
    idOcupacao: number;
    idProfissao: number;
    cep: string;
    logradouro: string;
    numero: string;
    ocupacao: string;
    renda: number;
    rendaAdicional: number;
    rendaTotal: number;
    telefoneDdd: string;
    telefoneNumero: string;
    telefoneRamal: string;
    // tempoAnos: number;
    // tempoMeses: number;
    cnpjEmpresa: string;   
    // dtAdmissao: Date;
    pessoaPolitica: boolean;
    pessoaRelacaoPolitica: boolean;
    
    ocupacaoPassivelAlteracao: boolean;
    empresaObrigatorio: boolean;
    enderecoObrigatorio: boolean;
    // dtAdmissaoObrigatorio: boolean;
    cnpjEmpresaObrigatorio: boolean;

    cargoAlterado: boolean;
    cidadeAlterado: boolean;
    bairroAlterado: boolean;
    complementoAlterado: boolean;
    dsProfissaoAlterado: boolean;
    empresaAlterado: boolean;
    idEstadoAlterado: boolean;
    estadoAlterado: boolean;
    idCargoAlterado: boolean;
    idOcupacaoAlterado: boolean;
    idProfissaoAlterado: boolean;
    cepAlterado: boolean;
    logradouroAlterado: boolean;
    numeroAlterado: boolean;
    ocupacaoAlterado: boolean;
    rendaAlterado: boolean;
    rendaAdicionalAlterado: boolean;
    rendaTotalAlterado: boolean;
    telefoneDddAlterado: boolean;
    telefoneNumeroAlterado: boolean;
    telefoneRamalAlterado: boolean;
    // tempoAnosAlterado: boolean;
    // tempoMesesAlterado: boolean;
    cnpjEmpresaAlterado: boolean;   
    // dtAdmissaoAlterado: boolean;
    pessoaPoliticaAlterado: boolean;
    pessoaRelacaoPoliticaAlterado: boolean;
}